Add class_deduction column to courses table if it does not exist
This commit introduces a new column 'class_deduction' to the 'courses' table in the 'testdb' database. The column is of type INT and has a default value of 1, with a comment indicating that it represents the number of classes to be deducted upon enrollment. The addition is conditional, ensuring that the column is only created if it does not already exist.